+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped minimum hikari version to 2.0.0.dev108.
 - Use `None` as the default for `max_value` and `min_value` in `parsing.py`.
 
+### Removed
+* `AbstractSchedule.iteration_count` (note, this property still exists on `IntervalSchedule`).
+
 ## [2.4.3a1] - 2022-03-12
 ### Added
 - Slash command "attachment" options.